Fixes for Merginho PR
