+++ src/core/search/search.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
-import { SearchService } from './search.service';
+import { SearchService } from './service/search.service';
 import { SearchController } from './search.controller';
 import { UsersModule } from '../users/users.module';
 import { envData } from 'src/config/datasource';
 import { OpensearchModule } from 'nestjs-opensearch';
+import { IndexService } from './service/create-index.service';
 
 @Module({
   imports: [
@@ -20,6 +21,6 @@ import { OpensearchModule } from 'nestjs-opensearch';
     }),
   ],
   controllers: [SearchController],
-  providers: [SearchService],
+  providers: [SearchService, IndexService],
 })
 export class SearchModule {}
